TRINITY: fixing a 64bit compile error
Binaries submitted in 1008801
Build time : 9 minutes 43.132000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1008750]

+++ Eve/SpaceObjectFactory/EveSOFDNA.cpp
@@ -251,7 +251,7 @@ void EveSOFDNA::SetupCustomData()
 			{
 				m_customHullData.additiveAreas.push_back( finder->second );
 				m_customHullData.additiveAreas[0].index = 0;
-				m_customHullData.additiveAreas[0].count = m_hullData->opaqueAreas.size();
+				m_customHullData.additiveAreas[0].count = (unsigned int)m_hullData->opaqueAreas.size();
 			}
 		}
 	}
